Split base SSE provider transformers
