+++ src/core/providers/base/sse.rs
@@ -5,16 +5,24 @@
 
 use bytes::Bytes;
 use futures::Stream;
-use serde_json::Value;
 use std::collections::VecDeque;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use tracing::warn;
 
-use crate::core::types::responses::{
-    ChatChunk, ChatDelta, ChatStreamChoice, FinishReason, FunctionCallDelta, ToolCallDelta,
-};
-use crate::core::{providers::unified_provider::ProviderError, types::thinking::ThinkingDelta};
+mod anthropic;
+mod cohere;
+mod databricks;
+mod gemini;
+mod openai;
+
+pub use anthropic::AnthropicTransformer;
+pub use cohere::CohereTransformer;
+pub use databricks::DatabricksTransformer;
+pub use gemini::GeminiTransformer;
+pub use openai::OpenAICompatibleTransformer;
+
+use crate::core::providers::unified_provider::ProviderError;
+use crate::core::types::responses::{ChatChunk, FinishReason};
 
 /// SSE Event Types
 #[derive(Debug, Clone, PartialEq)]
@@ -311,912 +319,6 @@ where
             Poll::Ready(None) => Poll::Ready(None),
             Poll::Pending => Poll::Pending,
         }
-    }
-}
-
-/// OpenAI-compatible SSE Transformer (can be reused by many providers)
-#[derive(Debug, Clone)]
-pub struct OpenAICompatibleTransformer {
-    provider: &'static str,
-}
-
-impl OpenAICompatibleTransformer {
-    pub fn new(provider: &'static str) -> Self {
-        Self { provider }
-    }
-}
-
-impl SSETransformer for OpenAICompatibleTransformer {
-    fn provider_name(&self) -> &'static str {
-        self.provider
-    }
-
-    fn transform_chunk(&self, data: &str) -> Result<Option<ChatChunk>, ProviderError> {
-        // Parse JSON
-        let json_value: Value = serde_json::from_str(data).map_err(|e| {
-            ProviderError::response_parsing(
-                self.provider,
-                format!("Failed to parse SSE JSON: {}", e),
-            )
-        })?;
-
-        // Extract fields
-        let id = json_value
-            .get("id")
-            .and_then(|v| v.as_str())
-            .unwrap_or("stream-chunk")
-            .to_string();
-
-        let model = json_value
-            .get("model")
-            .and_then(|v| v.as_str())
-            .unwrap_or("unknown")
-            .to_string();
-
-        let created = json_value
-            .get("created")
-            .and_then(|v| v.as_i64())
-            .unwrap_or_else(|| chrono::Utc::now().timestamp()) as u64;
-
-        // Parse choices
-        let choices = json_value
-            .get("choices")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| {
-                ProviderError::response_parsing(
-                    self.provider,
-                    "No choices in SSE chunk".to_string(),
-                )
-            })?;
-
-        let mut stream_choices = Vec::new();
-
-        for (index, choice) in choices.iter().enumerate() {
-            let delta = choice.get("delta").ok_or_else(|| {
-                ProviderError::response_parsing(self.provider, "No delta in choice".to_string())
-            })?;
-
-            let mut delta_obj: ChatDelta = serde_json::from_value(delta.clone()).map_err(|e| {
-                ProviderError::response_parsing(
-                    self.provider,
-                    format!("Failed to parse delta: {}", e),
-                )
-            })?;
-            if let Some(Value::String(reasoning_content)) = delta.get("reasoning_content") {
-                delta_obj.thinking = Some(ThinkingDelta {
-                    content: Some(reasoning_content.clone()),
-                    ..Default::default()
-                });
-            }
-
-            let finish_reason = choice
-                .get("finish_reason")
-                .and_then(|v| v.as_str())
-                .and_then(|s| self.parse_finish_reason(s));
-
-            let logprobs = choice
-                .get("logprobs")
-                .and_then(|v| serde_json::from_value(v.clone()).ok());
-
-            stream_choices.push(ChatStreamChoice {
-                index: index as u32,
-                delta: delta_obj,
-                finish_reason,
-                logprobs,
-            });
-        }
-
-        // Parse usage (optional)
-        let usage = json_value
-            .get("usage")
-            .and_then(|v| serde_json::from_value(v.clone()).ok());
-
-        Ok(Some(ChatChunk {
-            id,
-            object: "chat.completion.chunk".to_string(),
-            created: created as i64,
-            model,
-            choices: stream_choices,
-            usage,
-            system_fingerprint: json_value
-                .get("system_fingerprint")
-                .and_then(|v| v.as_str())
-                .map(|s| s.to_string()),
-        }))
-    }
-}
-
-/// Anthropic SSE Transformer
-///
-/// Handles Anthropic's event-based SSE format with message_start, content_block_delta,
-/// message_delta, and message_stop events.
-#[derive(Debug, Clone)]
-pub struct AnthropicTransformer {
-    model: String,
-}
-
-impl AnthropicTransformer {
-    pub fn new(model: impl Into<String>) -> Self {
-        Self {
-            model: model.into(),
-        }
-    }
-
-    fn parse_anthropic_finish_reason(reason: &str) -> FinishReason {
-        match reason {
-            "end_turn" => FinishReason::Stop,
-            "max_tokens" => FinishReason::Length,
-            "tool_use" => FinishReason::ToolCalls,
-            "stop_sequence" => FinishReason::StopSequence,
-            "refusal" => FinishReason::Refusal,
-            "pause_turn" => FinishReason::PauseTurn,
-            _ => FinishReason::Stop,
-        }
-    }
-
-    fn empty_delta() -> ChatDelta {
-        ChatDelta {
-            role: None,
-            content: None,
-            thinking: None,
-            tool_calls: None,
-            function_call: None,
-        }
-    }
-
-    fn chunk_with_choice(
-        &self,
-        created: i64,
-        delta: ChatDelta,
-        finish_reason: Option<FinishReason>,
-        usage: Option<crate::core::types::responses::Usage>,
-    ) -> ChatChunk {
-        ChatChunk {
-            id: String::new(),
-            object: "chat.completion.chunk".to_string(),
-            created,
-            model: self.model.clone(),
-            choices: vec![ChatStreamChoice {
-                index: 0,
-                delta,
-                finish_reason,
-                logprobs: None,
-            }],
-            usage,
-            system_fingerprint: None,
-        }
-    }
-}
-
-impl SSETransformer for AnthropicTransformer {
-    fn provider_name(&self) -> &'static str {
-        "anthropic"
-    }
-
-    fn transform_chunk(&self, data: &str) -> Result<Option<ChatChunk>, ProviderError> {
-        let json: Value = serde_json::from_str(data).map_err(|e| {
-            ProviderError::response_parsing(
-                "anthropic",
-                format!("Failed to parse Anthropic SSE: {}", e),
-            )
-        })?;
-
-        let event_type = json.get("type").and_then(|v| v.as_str()).unwrap_or("");
-
-        let created = chrono::Utc::now().timestamp();
-
-        match event_type {
-            "message_start" => {
-                let message_id = json
-                    .get("message")
-                    .and_then(|m| m.get("id"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("anthropic-stream")
-                    .to_string();
-
-                Ok(Some(ChatChunk {
-                    id: message_id,
-                    object: "chat.completion.chunk".to_string(),
-                    created,
-                    model: self.model.clone(),
-                    choices: vec![ChatStreamChoice {
-                        index: 0,
-                        delta: ChatDelta {
-                            role: Some(crate::core::types::message::MessageRole::Assistant),
-                            content: None,
-                            thinking: None,
-                            tool_calls: None,
-                            function_call: None,
-                        },
-                        finish_reason: None,
-                        logprobs: None,
-                    }],
-                    usage: None,
-                    system_fingerprint: None,
-                }))
-            }
-            "content_block_start" => {
-                let index = json.get("index").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-                let content_block = json.get("content_block").ok_or_else(|| {
-                    ProviderError::response_parsing(
-                        "anthropic",
-                        "No content_block in content_block_start".to_string(),
-                    )
-                })?;
-                let block_type = content_block
-                    .get("type")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-
-                match block_type {
-                    "tool_use" => {
-                        let id = content_block
-                            .get("id")
-                            .and_then(|v| v.as_str())
-                            .map(str::to_string);
-                        let name = content_block
-                            .get("name")
-                            .and_then(|v| v.as_str())
-                            .map(str::to_string);
-                        let arguments = content_block.get("input").and_then(|input| {
-                            if input.is_null()
-                                || input
-                                    .as_object()
-                                    .map(|object| object.is_empty())
-                                    .unwrap_or(false)
-                            {
-                                None
-                            } else {
-                                Some(input.to_string())
-                            }
-                        });
-
-                        let mut delta = Self::empty_delta();
-                        delta.tool_calls = Some(vec![ToolCallDelta {
-                            index,
-                            id,
-                            tool_type: Some("function".to_string()),
-                            function: Some(FunctionCallDelta { name, arguments }),
-                        }]);
-
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    "thinking" | "redacted_thinking" => {
-                        let mut delta = Self::empty_delta();
-                        delta.thinking = Some(ThinkingDelta::start());
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    "text" => Ok(None),
-                    _ => {
-                        warn!(
-                            provider = "anthropic",
-                            block_type, "Ignoring unknown Anthropic content block start"
-                        );
-                        Ok(None)
-                    }
-                }
-            }
-            "content_block_delta" => {
-                let index = json.get("index").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-                let delta_json = json.get("delta").ok_or_else(|| {
-                    ProviderError::response_parsing(
-                        "anthropic",
-                        "No delta in content_block_delta".to_string(),
-                    )
-                })?;
-                let delta_type = delta_json
-                    .get("type")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-
-                match delta_type {
-                    "text_delta" => {
-                        let text = delta_json
-                            .get("text")
-                            .and_then(|t| t.as_str())
-                            .unwrap_or("");
-                        let mut delta = Self::empty_delta();
-                        delta.content = Some(text.to_string());
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    "input_json_delta" => {
-                        let partial_json = delta_json
-                            .get("partial_json")
-                            .and_then(|value| value.as_str())
-                            .unwrap_or("");
-                        let mut delta = Self::empty_delta();
-                        delta.tool_calls = Some(vec![ToolCallDelta {
-                            index,
-                            id: None,
-                            tool_type: Some("function".to_string()),
-                            function: Some(FunctionCallDelta {
-                                name: None,
-                                arguments: Some(partial_json.to_string()),
-                            }),
-                        }]);
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    "thinking_delta" => {
-                        let thinking = delta_json
-                            .get("thinking")
-                            .and_then(|value| value.as_str())
-                            .unwrap_or("");
-                        let mut delta = Self::empty_delta();
-                        delta.thinking = Some(ThinkingDelta::new(thinking));
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    "signature_delta" => {
-                        let signature = delta_json
-                            .get("signature")
-                            .and_then(|value| value.as_str())
-                            .unwrap_or("");
-                        let mut delta = Self::empty_delta();
-                        delta.thinking = Some(ThinkingDelta {
-                            signature: Some(signature.to_string()),
-                            ..Default::default()
-                        });
-                        Ok(Some(self.chunk_with_choice(created, delta, None, None)))
-                    }
-                    _ => {
-                        warn!(
-                            provider = "anthropic",
-                            delta_type, "Ignoring unknown Anthropic content block delta"
-                        );
-                        Ok(None)
-                    }
-                }
-            }
-            "message_delta" => {
-                let finish_reason = json
-                    .get("delta")
-                    .and_then(|d| d.get("stop_reason"))
-                    .and_then(|r| r.as_str())
-                    .map(Self::parse_anthropic_finish_reason);
-
-                let usage = json.get("usage").map(|u| {
-                    let input = u.get("input_tokens").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-                    let output =
-                        u.get("output_tokens").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-                    crate::core::types::responses::Usage {
-                        prompt_tokens: input,
-                        completion_tokens: output,
-                        total_tokens: input + output,
-                        completion_tokens_details: None,
-                        prompt_tokens_details: None,
-                        thinking_usage: None,
-                    }
-                });
-
-                Ok(Some(self.chunk_with_choice(
-                    created,
-                    Self::empty_delta(),
-                    finish_reason,
-                    usage,
-                )))
-            }
-            "message_stop" => Ok(Some(ChatChunk {
-                id: String::new(),
-                object: "chat.completion.chunk".to_string(),
-                created,
-                model: self.model.clone(),
-                choices: vec![],
-                usage: None,
-                system_fingerprint: None,
-            })),
-            "error" => {
-                let msg = json
-                    .get("error")
-                    .and_then(|e| e.get("message"))
-                    .and_then(|m| m.as_str())
-                    .unwrap_or("Unknown streaming error");
-                Err(ProviderError::streaming_error(
-                    "anthropic",
-                    "chat",
-                    None,
-                    None,
-                    msg.to_string(),
-                ))
-            }
-            "content_block_stop" | "ping" => Ok(None),
-            _ => {
-                warn!(
-                    provider = "anthropic",
-                    event_type, "Ignoring unknown Anthropic SSE event type"
-                );
-                Ok(None)
-            }
-        }
-    }
-}
-
-/// Gemini SSE Transformer
-///
-/// Handles Gemini's streaming format with candidates/parts structure.
-#[derive(Debug, Clone)]
-pub struct GeminiTransformer {
-    model: String,
-    chunk_id: String,
-}
-
-impl GeminiTransformer {
-    pub fn new(model: impl Into<String>) -> Self {
-        let nanos = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_nanos())
-            .unwrap_or(0);
-        Self {
-            model: model.into(),
-            chunk_id: format!("gemini-stream-{}", nanos),
-        }
-    }
-}
-
-impl SSETransformer for GeminiTransformer {
-    fn provider_name(&self) -> &'static str {
-        "gemini"
-    }
-
-    fn transform_chunk(&self, data: &str) -> Result<Option<ChatChunk>, ProviderError> {
-        let json: Value = serde_json::from_str(data).map_err(|e| {
-            ProviderError::response_parsing("gemini", format!("Failed to parse Gemini SSE: {}", e))
-        })?;
-
-        // Error response
-        if let Some(error) = json.get("error") {
-            let msg = error
-                .get("message")
-                .and_then(|m| m.as_str())
-                .unwrap_or("Unknown Gemini error");
-            return Err(ProviderError::api_error(
-                "gemini",
-                error.get("code").and_then(|c| c.as_u64()).unwrap_or(500) as u16,
-                msg.to_string(),
-            ));
-        }
-
-        let empty_arr = vec![];
-        let candidates = json
-            .get("candidates")
-            .and_then(|c| c.as_array())
-            .unwrap_or(&empty_arr);
-
-        if candidates.is_empty() {
-            // Usage-only chunk
-            let usage = json.get("usageMetadata").map(|u| {
-                let prompt = u
-                    .get("promptTokenCount")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0) as u32;
-                let completion = u
-                    .get("candidatesTokenCount")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0) as u32;
-                let total = u
-                    .get("totalTokenCount")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0) as u32;
-                crate::core::types::responses::Usage {
-                    prompt_tokens: prompt,
-                    completion_tokens: completion,
-                    total_tokens: total,
-                    prompt_tokens_details: None,
-                    completion_tokens_details: None,
-                    thinking_usage: None,
-                }
-            });
-            if usage.is_some() {
-                return Ok(Some(ChatChunk {
-                    id: self.chunk_id.clone(),
-                    object: "chat.completion.chunk".to_string(),
-                    created: chrono::Utc::now().timestamp(),
-                    model: self.model.clone(),
-                    choices: vec![],
-                    usage,
-                    system_fingerprint: None,
-                }));
-            }
-            return Ok(None);
-        }
-
-        let mut choices = Vec::new();
-        for (index, candidate) in candidates.iter().enumerate() {
-            let empty_parts = vec![];
-            let parts = candidate
-                .get("content")
-                .and_then(|c| c.get("parts"))
-                .and_then(|p| p.as_array())
-                .unwrap_or(&empty_parts);
-
-            let mut text_parts = Vec::new();
-            for part in parts {
-                if let Some(text) = part.get("text").and_then(|t| t.as_str()) {
-                    text_parts.push(text);
-                }
-            }
-            let delta_content = text_parts.join("");
-
-            let finish_reason = candidate
-                .get("finishReason")
-                .and_then(|r| r.as_str())
-                .map(|r| match r {
-                    "STOP" => FinishReason::Stop,
-                    "MAX_TOKENS" => FinishReason::Length,
-                    "SAFETY" | "RECITATION" => FinishReason::ContentFilter,
-                    _ => FinishReason::Stop,
-                });
-
-            choices.push(ChatStreamChoice {
-                index: index as u32,
-                delta: ChatDelta {
-                    role: if !delta_content.is_empty() || finish_reason.is_some() {
-                        Some(crate::core::types::message::MessageRole::Assistant)
-                    } else {
-                        None
-                    },
-                    content: if delta_content.is_empty() {
-                        None
-                    } else {
-                        Some(delta_content)
-                    },
-                    thinking: None,
-                    function_call: None,
-                    tool_calls: None,
-                },
-                finish_reason,
-                logprobs: None,
-            });
-        }
-
-        let usage = json.get("usageMetadata").map(|u| {
-            let prompt = u
-                .get("promptTokenCount")
-                .and_then(|v| v.as_u64())
-                .unwrap_or(0) as u32;
-            let completion = u
-                .get("candidatesTokenCount")
-                .and_then(|v| v.as_u64())
-                .unwrap_or(0) as u32;
-            let total = u
-                .get("totalTokenCount")
-                .and_then(|v| v.as_u64())
-                .unwrap_or(0) as u32;
-            crate::core::types::responses::Usage {
-                prompt_tokens: prompt,
-                completion_tokens: completion,
-                total_tokens: total,
-                prompt_tokens_details: None,
-                completion_tokens_details: None,
-                thinking_usage: None,
-            }
-        });
-
-        if choices.is_empty() && usage.is_none() {
-            return Ok(None);
-        }
-
-        Ok(Some(ChatChunk {
-            id: self.chunk_id.clone(),
-            object: "chat.completion.chunk".to_string(),
-            created: chrono::Utc::now().timestamp(),
-            model: self.model.clone(),
-            choices,
-            usage,
-            system_fingerprint: None,
-        }))
-    }
-}
-
-/// Cohere SSE Transformer
-///
-/// Handles Cohere's streaming format with v1/v2 API version support.
-#[derive(Debug, Clone)]
-pub struct CohereTransformer {
-    model: String,
-    response_id: String,
-    /// true = v2, false = v1
-    use_v2: bool,
-}
-
-impl CohereTransformer {
-    pub fn new(model: impl Into<String>, use_v2: bool) -> Self {
-        Self {
-            model: model.into(),
-            response_id: format!("chatcmpl-{}", uuid::Uuid::new_v4()),
-            use_v2,
-        }
-    }
-
-    fn parse_cohere_finish_reason(reason: &str) -> FinishReason {
-        match reason.to_lowercase().as_str() {
-            "stop" | "complete" | "end_turn" => FinishReason::Stop,
-            "length" | "max_tokens" => FinishReason::Length,
-            "tool_calls" | "tool_use" => FinishReason::ToolCalls,
-            "content_filter" => FinishReason::ContentFilter,
-            _ => FinishReason::Stop,
-        }
-    }
-}
-
-impl SSETransformer for CohereTransformer {
-    fn provider_name(&self) -> &'static str {
-        "cohere"
-    }
-
-    fn transform_chunk(&self, data: &str) -> Result<Option<ChatChunk>, ProviderError> {
-        let json: Value = serde_json::from_str(data).map_err(|e| {
-            ProviderError::response_parsing("cohere", format!("Failed to parse Cohere SSE: {}", e))
-        })?;
-
-        let event_type = json
-            .get("type")
-            .or_else(|| json.get("event"))
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-
-        let created = chrono::Utc::now().timestamp();
-
-        if self.use_v2 {
-            match event_type {
-                "content-delta" => {
-                    let text = json
-                        .get("delta")
-                        .and_then(|d| d.get("message"))
-                        .and_then(|m| m.get("content"))
-                        .and_then(|c| {
-                            c.get("text")
-                                .and_then(|t| t.as_str())
-                                .or_else(|| c.as_str())
-                        })
-                        .unwrap_or("");
-
-                    if text.is_empty() {
-                        return Ok(None);
-                    }
-
-                    Ok(Some(ChatChunk {
-                        id: self.response_id.clone(),
-                        object: "chat.completion.chunk".to_string(),
-                        created,
-                        model: self.model.clone(),
-                        choices: vec![ChatStreamChoice {
-                            index: 0,
-                            delta: ChatDelta {
-                                role: None,
-                                content: Some(text.to_string()),
-                                thinking: None,
-                                tool_calls: None,
-                                function_call: None,
-                            },
-                            finish_reason: None,
-                            logprobs: None,
-                        }],
-                        usage: None,
-                        system_fingerprint: None,
-                    }))
-                }
-                "message-end" => {
-                    let data_field = json.get("data").or(json.get("delta"));
-                    let finish_reason = data_field
-                        .and_then(|d| d.get("delta"))
-                        .and_then(|d| d.get("finish_reason"))
-                        .and_then(|f| f.as_str())
-                        .unwrap_or("stop");
-
-                    let usage = data_field
-                        .and_then(|d| d.get("delta"))
-                        .and_then(|d| d.get("usage"))
-                        .and_then(|u| u.get("tokens"))
-                        .map(|tokens| {
-                            let prompt = tokens
-                                .get("input_tokens")
-                                .and_then(|v| v.as_u64())
-                                .unwrap_or(0) as u32;
-                            let completion = tokens
-                                .get("output_tokens")
-                                .and_then(|v| v.as_u64())
-                                .unwrap_or(0) as u32;
-                            crate::core::types::responses::Usage {
-                                prompt_tokens: prompt,
-                                completion_tokens: completion,
-                                total_tokens: prompt + completion,
-                                prompt_tokens_details: None,
-                                completion_tokens_details: None,
-                                thinking_usage: None,
-                            }
-                        });
-
-                    Ok(Some(ChatChunk {
-                        id: self.response_id.clone(),
-                        object: "chat.completion.chunk".to_string(),
-                        created,
-                        model: self.model.clone(),
-                        choices: vec![ChatStreamChoice {
-                            index: 0,
-                            delta: ChatDelta {
-                                role: None,
-                                content: None,
-                                thinking: None,
-                                tool_calls: None,
-                                function_call: None,
-                            },
-                            finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
-                            logprobs: None,
-                        }],
-                        usage,
-                        system_fingerprint: None,
-                    }))
-                }
-                // message-start, content-start, content-end, tool-call-*, citation-* — skip
-                _ => Ok(None),
-            }
-        } else {
-            // v1 format
-            match event_type {
-                "text-generation" => {
-                    let text = json.get("text").and_then(|t| t.as_str()).unwrap_or("");
-                    Ok(Some(ChatChunk {
-                        id: self.response_id.clone(),
-                        object: "chat.completion.chunk".to_string(),
-                        created,
-                        model: self.model.clone(),
-                        choices: vec![ChatStreamChoice {
-                            index: 0,
-                            delta: ChatDelta {
-                                role: None,
-                                content: Some(text.to_string()),
-                                thinking: None,
-                                tool_calls: None,
-                                function_call: None,
-                            },
-                            finish_reason: None,
-                            logprobs: None,
-                        }],
-                        usage: None,
-                        system_fingerprint: None,
-                    }))
-                }
-                "stream-end" => {
-                    let finish_reason = json
-                        .get("finish_reason")
-                        .and_then(|f| f.as_str())
-                        .unwrap_or("stop");
-                    Ok(Some(ChatChunk {
-                        id: self.response_id.clone(),
-                        object: "chat.completion.chunk".to_string(),
-                        created,
-                        model: self.model.clone(),
-                        choices: vec![ChatStreamChoice {
-                            index: 0,
-                            delta: ChatDelta {
-                                role: None,
-                                content: None,
-                                thinking: None,
-                                tool_calls: None,
-                                function_call: None,
-                            },
-                            finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
-                            logprobs: None,
-                        }],
-                        usage: None,
-                        system_fingerprint: None,
-                    }))
-                }
-                // stream-start, citation-generation, tool-calls-generation — skip
-                _ => Ok(None),
-            }
-        }
-    }
-}
-
-/// Databricks SSE Transformer
-///
-/// OpenAI-compatible format with additional support for array content (Claude-style).
-#[derive(Debug, Clone)]
-pub struct DatabricksTransformer;
-
-impl SSETransformer for DatabricksTransformer {
-    fn provider_name(&self) -> &'static str {
-        "databricks"
-    }
-
-    fn transform_chunk(&self, data: &str) -> Result<Option<ChatChunk>, ProviderError> {
-        let json: Value = serde_json::from_str(data).map_err(|e| {
-            ProviderError::response_parsing(
-                "databricks",
-                format!("Failed to parse SSE JSON: {}", e),
-            )
-        })?;
-
-        let id = json
-            .get("id")
-            .and_then(|v| v.as_str())
-            .unwrap_or("chunk")
-            .to_string();
-        let created = json
-            .get("created")
-            .and_then(|v| v.as_i64())
-            .unwrap_or_else(|| chrono::Utc::now().timestamp());
-        let model = json
-            .get("model")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-
-        let mut choices = Vec::new();
-        if let Some(choices_arr) = json.get("choices").and_then(|v| v.as_array()) {
-            for choice in choices_arr {
-                let index = choice.get("index").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-
-                let delta = if let Some(delta_obj) = choice.get("delta") {
-                    let role =
-                        delta_obj
-                            .get("role")
-                            .and_then(|v| v.as_str())
-                            .and_then(|r| match r {
-                                "assistant" => {
-                                    Some(crate::core::types::message::MessageRole::Assistant)
-                                }
-                                "user" => Some(crate::core::types::message::MessageRole::User),
-                                "system" => Some(crate::core::types::message::MessageRole::System),
-                                "tool" => Some(crate::core::types::message::MessageRole::Tool),
-                                _ => None,
-                            });
-
-                    // Handle content — could be string or array (Claude reasoning)
-                    let content = match delta_obj.get("content") {
-                        Some(Value::String(s)) => Some(s.clone()),
-                        Some(Value::Array(arr)) => {
-                            let mut text = String::new();
-                            for item in arr {
-                                if let Some(t) = item.get("text").and_then(|v| v.as_str()) {
-                                    text.push_str(t);
-                                }
-                            }
-                            if text.is_empty() { None } else { Some(text) }
-                        }
-                        _ => None,
-                    };
-
-                    ChatDelta {
-                        role,
-                        content,
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                    }
-                } else {
-                    ChatDelta {
-                        role: None,
-                        content: None,
-                        thinking: None,
-                        tool_calls: None,
-                        function_call: None,
-                    }
-                };
-
-                let finish_reason = choice
-                    .get("finish_reason")
-                    .and_then(|v| v.as_str())
-                    .and_then(|r| self.parse_finish_reason(r));
-
-                choices.push(ChatStreamChoice {
-                    index,
-                    delta,
-                    finish_reason,
-                    logprobs: None,
-                });
-            }
-        }
-
-        Ok(Some(ChatChunk {
-            id,
-            object: "chat.completion.chunk".to_string(),
-            created,
-            model,
-            choices,
-            usage: None,
-            system_fingerprint: None,
-        }))
     }
 }
